core: api-server: http: Refactor handlers into mod for discoverability
