+++ core/src/api_server/http.rs
@@ -1,11 +1,10 @@
 //! Groups handlers for the HTTP API
 
 use async_trait::async_trait;
-use crossbeam::channel;
 use hyper::{
     server::conn::AddrStream,
     service::{make_service_fn, service_fn},
-    Body, Error as HyperError, Method, Request, Response, Server, StatusCode,
+    Body, Error as HyperError, Method, Request, Response, Server,
 };
 use std::{
     convert::Infallible,
@@ -13,19 +12,18 @@ use std::{
     sync::Arc,
     time::{SystemTime, UNIX_EPOCH},
 };
-use uuid::Uuid;
 
 use crate::{
-    external_api::{
-        http::{
-            GetExchangeHealthStatesRequest, GetExchangeHealthStatesResponse, GetOrderByIdResponse,
-            GetOrdersResponse, GetWalletResponse, PingResponse,
-        },
-        types::Wallet,
-        EmptyRequestResponse,
-    },
-    price_reporter::jobs::PriceReporterManagerJob,
+    external_api::{http::PingResponse, EmptyRequestResponse},
     state::RelayerState,
+};
+
+use self::{
+    price_report::{ExchangeHealthStatesHandler, EXCHANGE_HEALTH_ROUTE},
+    wallet::{
+        GetOrderByIdHandler, GetOrdersHandler, GetWalletHandler, GET_ORDERS_ROUTE,
+        GET_ORDER_BY_ID_ROUTE, GET_WALLET_ROUTE,
+    },
 };
 
 use super::{
@@ -34,38 +32,11 @@ use super::{
     worker::ApiServerConfig,
 };
 
-/// The :wallet_id param in a URL
-const WALLET_ID_URL_PARAM: &str = "wallet_id";
-/// The :order_id param in a URL
-const ORDER_ID_URL_PARAM: &str = "order_id";
-
-/// Error message displayed when a given order ID is not parsable
-const ERR_ORDER_ID_PARSE: &str = "could not parse order id";
-/// Error message displayed when a given wallet ID is not parsable
-const ERR_WALLET_ID_PARSE: &str = "could not parse wallet id";
-/// Error message displayed when a given order cannot be found
-const ERR_ORDER_NOT_FOUND: &str = "order not found";
-/// The error message to display when a wallet cannot be found
-const ERR_WALLET_NOT_FOUND: &str = "wallet not found";
-
-// ---------------
-// | HTTP Routes |
-// ---------------
+mod price_report;
+mod wallet;
 
 /// Health check
 const PING_ROUTE: &str = "/v0/ping";
-/// Exchange health check route
-const EXCHANGE_HEALTH_ROUTE: &str = "/v0/exchange/health_check";
-/// Returns the wallet information for the given id
-const GET_WALLET_ROUTE: &str = "/v0/wallet/:wallet_id";
-/// Returns the orders within a given wallet
-const GET_ORDERS_ROUTE: &str = "/v0/wallet/:wallet_id/orders";
-/// Returns a single order by the given identifier
-const GET_ORDER_BY_ID_ROUTE: &str = "/v0/wallet/:wallet_id/orders/:order_id";
-
-// ----------------
-// | Router Setup |
-// ----------------
 
 /// A wrapper around the router and task management operations that
 /// the worker may delegate to
@@ -163,10 +134,6 @@ impl HttpServer {
     }
 }
 
-// ----------------
-// | Generic APIs |
-// ----------------
-
 /// Handler for the ping route, returns a pong
 #[derive(Clone, Debug)]
 pub struct PingHandler;
@@ -192,242 +159,5 @@ impl TypedHandler for PingHandler {
             .unwrap()
             .as_millis();
         Ok(PingResponse { timestamp })
-    }
-}
-
-// --------------------------
-// | Wallet Operations APIs |
-// --------------------------
-
-/// Handler for the GET /wallet/:id route
-#[derive(Debug)]
-pub struct GetWalletHandler {
-    /// A copy of the relayer-global state
-    global_state: RelayerState,
-}
-
-impl GetWalletHandler {
-    /// Create a new handler for the /v0/wallet/:id route
-    pub fn new(global_state: RelayerState) -> Self {
-        Self { global_state }
-    }
-}
-
-#[async_trait]
-impl TypedHandler for GetWalletHandler {
-    type Request = EmptyRequestResponse;
-    type Response = GetWalletResponse;
-
-    async fn handle_typed(
-        &self,
-        _req: Self::Request,
-        params: UrlParams,
-    ) -> Result<Self::Response, ApiServerError> {
-        let wallet_id: Uuid = params
-            .get(WALLET_ID_URL_PARAM)
-            .unwrap()
-            .parse()
-            .map_err(|_| {
-                ApiServerError::HttpStatusCode(
-                    StatusCode::BAD_REQUEST,
-                    ERR_WALLET_ID_PARSE.to_string(),
-                )
-            })?;
-
-        if let Some(wallet) = self
-            .global_state
-            .read_wallet_index()
-            .await
-            .get_wallet(&wallet_id)
-            .await
-        {
-            Ok(GetWalletResponse {
-                wallet: wallet.into(),
-            })
-        } else {
-            Err(ApiServerError::HttpStatusCode(
-                StatusCode::NOT_FOUND,
-                ERR_WALLET_NOT_FOUND.to_string(),
-            ))
-        }
-    }
-}
-
-/// Handler for the GET /wallet/:id/orders route
-#[derive(Clone, Debug)]
-pub struct GetOrdersHandler {
-    /// A copy of the relayer-global state
-    pub global_state: RelayerState,
-}
-
-impl GetOrdersHandler {
-    /// Create a new handler for the /wallet/:id/orders route
-    pub fn new(global_state: RelayerState) -> Self {
-        Self { global_state }
-    }
-}
-
-#[async_trait]
-impl TypedHandler for GetOrdersHandler {
-    type Request = EmptyRequestResponse;
-    type Response = GetOrdersResponse;
-
-    async fn handle_typed(
-        &self,
-        _req: Self::Request,
-        params: UrlParams,
-    ) -> Result<Self::Response, ApiServerError> {
-        let wallet_id: Uuid = params
-            .get(WALLET_ID_URL_PARAM)
-            .unwrap()
-            .parse()
-            .map_err(|_| {
-                ApiServerError::HttpStatusCode(
-                    StatusCode::BAD_REQUEST,
-                    ERR_WALLET_NOT_FOUND.to_string(),
-                )
-            })?;
-
-        if let Some(wallet) = self
-            .global_state
-            .read_wallet_index()
-            .await
-            .get_wallet(&wallet_id)
-            .await
-        {
-            let wallet: Wallet = wallet.into();
-            Ok(GetOrdersResponse {
-                orders: wallet.orders,
-            })
-        } else {
-            Err(ApiServerError::HttpStatusCode(
-                StatusCode::NOT_FOUND,
-                ERR_WALLET_NOT_FOUND.to_string(),
-            ))
-        }
-    }
-}
-
-/// Handler for the GET /wallet/:id/orders/:id route
-#[derive(Clone, Debug)]
-pub struct GetOrderByIdHandler {
-    /// A copy of the relayer-global state
-    pub global_state: RelayerState,
-}
-
-impl GetOrderByIdHandler {
-    /// Constructor
-    pub fn new(global_state: RelayerState) -> Self {
-        Self { global_state }
-    }
-}
-
-#[async_trait]
-impl TypedHandler for GetOrderByIdHandler {
-    type Request = EmptyRequestResponse;
-    type Response = GetOrderByIdResponse;
-
-    async fn handle_typed(
-        &self,
-        _req: Self::Request,
-        params: UrlParams,
-    ) -> Result<Self::Response, ApiServerError> {
-        let wallet_id: Uuid = params
-            .get(WALLET_ID_URL_PARAM)
-            .unwrap()
-            .parse()
-            .map_err(|_| {
-                ApiServerError::HttpStatusCode(
-                    StatusCode::BAD_REQUEST,
-                    ERR_WALLET_ID_PARSE.to_string(),
-                )
-            })?;
-        let order_id: Uuid = params
-            .get(ORDER_ID_URL_PARAM)
-            .unwrap()
-            .parse()
-            .map_err(|_| {
-                ApiServerError::HttpStatusCode(
-                    StatusCode::BAD_REQUEST,
-                    ERR_ORDER_ID_PARSE.to_string(),
-                )
-            })?;
-
-        if let Some(order) = (|| async {
-            self.global_state
-                .read_wallet_index()
-                .await
-                .get_wallet(&wallet_id)
-                .await?
-                .orders
-                .get(&order_id)
-                .cloned()
-        })()
-        .await
-        {
-            Ok(GetOrderByIdResponse {
-                order: (order_id, order).into(),
-            })
-        } else {
-            Err(ApiServerError::HttpStatusCode(
-                StatusCode::NOT_FOUND,
-                ERR_ORDER_NOT_FOUND.to_string(),
-            ))
-        }
-    }
-}
-
-// ------------------------
-// | Price Reporting APIs |
-// ------------------------
-
-/// Handler for the / route, returns the health report for each individual
-/// exchange and the aggregate median
-#[derive(Clone, Debug)]
-pub(crate) struct ExchangeHealthStatesHandler {
-    /// The config for the API server
-    config: ApiServerConfig,
-}
-
-impl ExchangeHealthStatesHandler {
-    /// Create a new handler for "/exchange/health"
-    pub fn new(config: ApiServerConfig) -> Self {
-        Self { config }
-    }
-}
-
-#[async_trait]
-impl TypedHandler for ExchangeHealthStatesHandler {
-    type Request = GetExchangeHealthStatesRequest;
-    type Response = GetExchangeHealthStatesResponse;
-
-    async fn handle_typed(
-        &self,
-        req: Self::Request,
-        _params: UrlParams,
-    ) -> Result<Self::Response, ApiServerError> {
-        let (price_reporter_state_sender, price_reporter_state_receiver) = channel::unbounded();
-        self.config
-            .price_reporter_work_queue
-            .send(PriceReporterManagerJob::PeekMedian {
-                base_token: req.base_token.clone(),
-                quote_token: req.quote_token.clone(),
-                channel: price_reporter_state_sender,
-            })
-            .unwrap();
-        let (exchange_connection_state_sender, exchange_connection_state_receiver) =
-            channel::unbounded();
-        self.config
-            .price_reporter_work_queue
-            .send(PriceReporterManagerJob::PeekAllExchanges {
-                base_token: req.base_token,
-                quote_token: req.quote_token,
-                channel: exchange_connection_state_sender,
-            })
-            .unwrap();
-        Ok(GetExchangeHealthStatesResponse {
-            median: price_reporter_state_receiver.recv().unwrap(),
-            all_exchanges: exchange_connection_state_receiver.recv().unwrap(),
-        })
     }
 }
