+++ core/src/external_api/http/mod.rs
@@ -1,0 +1,13 @@
+//! Groups API types for the HTTP API
+
+use serde::{Deserialize, Serialize};
+
+pub mod price_report;
+pub mod wallet;
+
+/// A ping response
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct PingResponse {
+    /// The timestamp when the response is sent
+    pub timestamp: u128,
+}
